Dynamically (un)subscribe to topics if output is requested. Solving issue #52
Implementing a `SubscriberStatusCallback` which is triggered every time
someone (un)subscribes to the published topic, using it to (un)subscribe
to topics that supply the data for the actual calculations.
This allows to have the people perception running without using any CPU
if there is no component connected to it and only do calculations if another
node connects to it.