Fixing an issue with one of the association algorithms in the bayestracking library
The bayestracking library offered three assiciation algorithms: NN, JPDA and NNJPDA.
JPDA however was not defined in the code any more and is just a relic from past development processes.
The README and config file now reflect that only NN and NNJPDA are available.
Also, the people tracker is now checking if a valid alogirthm was defined in the config file instead of just using NNJPDA if it is not NN or JPDA.