Make the linker happy.
Longer explanation: the mongodb_store links to static libmongoclient.a which needs dynamic libboost-filesystem-mt.so.
The problem is another component links to that boost one too, earlier. So CMake doesn't add that into the linker line
later when mongodb_store kicks in. Moving it to the front fixes it, even though it's not clean.