Making camera topic reconfigurable
So far only the camera namespace was configurable but that introduced an implicit dependency on the openni_wrapper.
With these changes the whole topic is reconfigurable via a parameter, e.g.:

camera_namespace:=/my_cam
depth_image:=/depth/image

results in `/my_cam/depth/image` as a topic for the depth image. So `camera_namespace` + `depth_image` = the topic on which to look for the depth image.